General Election to : Results June-2024
The Election Commission of India (ECI) announced the schedule for the 2024 General Elections to the 18th Lok Sabha and the Legislative Assemblies of Andhra Pradesh, Arunachal Pradesh, Odisha, and Sikkim on March 16, 2024. The elections, including certain bye-elections, were conducted in seven phases, ensuring a smooth and comprehensive electoral process across the vast and diverse country.

Phase 1 of the elections commenced on April 19, 2024, with a voter turnout of 66.14%. The turnout slightly increased to 66.71% in Phase 2, held on April 26, 2024. Phase 3, which took place on May 7, 2024, saw a turnout of 65.68%. Notably, Phase 4, conducted on May 13, 2024, recorded the highest voter turnout at 69.16%, demonstrating peak participation. The turnout dipped in Phase 5, held on May 20, 2024, to 62.2%. Phase 6 on May 25, 2024, had a turnout of 63.37%, while Phase 7, the final phase held on June 1, 2024, recorded a turnout of 61.63%.